Fix unit tests where "only number" was causing issue.
For some reason, Bob.swift was in the repo, but not in the
Xcode project, so I added it into Xcode.

I found a method called "countElements(...)" that will return
a count, for us we needed length of a string, and in this case
it returns the length of the string "input". From there, the
regex we worked on would grab the correct range, and so I
modified the return statement to see if matches.count != 0 (true),
or false if 0.

Also added a .playground file so we can see tests in action,
but also to illustrate different test cases for our functions we
are writing in our tests.

One last change was that there is a .hasSuffix(...) method
on a String instance, so I adjusted the isQuestion(input:)
method to use that, rather than the ternary operator.