Complete tests and Bob class
I was on a roll and wanted to finish them up, and figured out the last two
for the "silent" tests. I first tested different scenarios in the Playground
(which is SUPER helpful), and then derived that my tests should work,
and they passed. Thought I would post for everyone to see.

Added isSilent method to return whether no string or a series of blank
characters was added. Nice thing in Swift is that you can iterate in a
for-in block over each character in a String object, so that made it easy.
I toyed with expanding the RegEx at first, but that just made it harder.